plugins: rename schema to schemas
Why:

* comply with wazo-notebook/file-structure.md
