+++ wazo_calld/plugins/applications/schemas.py
