+++ wazo_calld/plugins/endpoints/http.py
@@ -6,7 +6,7 @@ from xivo.tenant_flask_helpers import Tenant
 from wazo_calld.auth import required_acl
 from wazo_calld.http import AuthResource
 
-from .schema import trunk_endpoint_schema
+from .schemas import trunk_endpoint_schema
 
 
 class TrunkEndpoints(AuthResource):
