+++ wazo_calld/plugins/endpoints/schemas.py
