+++ wazo_calld/plugins/relocates/schemas.py
